// src/commons/types/generated/types.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>
}
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>
}
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any
  /** The `Upload` scalar type represents a file upload. */
  Upload: any
}

export type IBoard = {
  __typename?: 'Board'
  _id: Scalars['ID']
  boardAddress?: Maybe<IBoardAddress>
  contents: Scalars['String']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  dislikeCount: Scalars['Int']
  images?: Maybe<Array<Scalars['String']>>
  likeCount: Scalars['Int']
  title: Scalars['String']
  updatedAt: Scalars['DateTime']
  user?: Maybe<IUser>
  writer?: Maybe<Scalars['String']>
  youtubeUrl?: Maybe<Scalars['String']>
}

export type IBoardAddress = {
  __typename?: 'BoardAddress'
  _id: Scalars['ID']
  address?: Maybe<Scalars['String']>
  addressDetail?: Maybe<Scalars['String']>
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  updatedAt: Scalars['DateTime']
  zipcode?: Maybe<Scalars['String']>
}

export type IBoardAddressInput = {
  address?: InputMaybe<Scalars['String']>
  addressDetail?: InputMaybe<Scalars['String']>
  zipcode?: InputMaybe<Scalars['String']>
}

export type IBoardComment = {
  __typename?: 'BoardComment'
  _id: Scalars['ID']
  contents: Scalars['String']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  rating: Scalars['Float']
  updatedAt: Scalars['DateTime']
  user?: Maybe<IUser>
  writer?: Maybe<Scalars['String']>
}
// ...
export type IFileManager = {
  __typename?: 'FileManager'
  _id: Scalars['ID']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  isUsed: Scalars['Boolean']
  size?: Maybe<Scalars['Float']>
  updatedAt: Scalars['DateTime']
  url: Scalars['String']
}

export type IMutation = {
  __typename?: 'Mutation'
  createBoard: IBoard
  createBoardComment: IBoardComment
  createPointTransactionOfBuyingAndSelling: IUseditem
  createPointTransactionOfLoading: IPointTransaction
  createUseditem: IUseditem
  createUseditemQuestion: IUseditemQuestion
  createUseditemQuestionAnswer: IUseditemQuestionAnswer
  createUser: IUser
  deleteBoard: Scalars['ID']
  deleteBoardComment: Scalars['ID']
  deleteBoards: Array<Scalars['ID']>
  deleteUseditem: Scalars['ID']
  deleteUseditemQuestion: Scalars['ID']
  deleteUseditemQuestionAnswer: Scalars['String']
  dislikeBoard: Scalars['Int']
  likeBoard: Scalars['Int']
  loginUser: IToken
  loginUserExample: IToken
  logoutUser: Scalars['Boolean']
  resetUserPassword: Scalars['Boolean']
  restoreAccessToken: IToken
  toggleUseditemPick: Scalars['Int']
  updateBoard: IBoard
  updateBoardComment: IBoardComment
  updateUseditem: IUseditem
  updateUseditemQuestion: IUseditemQuestion
  updateUseditemQuestionAnswer: IUseditemQuestionAnswer
  updateUser: IUser
  uploadFile: IFileManager
}
// ...
export type IMutationDeleteBoardArgs = {
  boardId: Scalars['ID']
}
// ...
export type IMutationDislikeBoardArgs = {
  boardId: Scalars['ID']
}

export type IMutationLikeBoardArgs = {
  boardId: Scalars['ID']
}
// ...
export type IPointTransaction = {
  __typename?: 'PointTransaction'
  _id: Scalars['ID']
  amount: Scalars['Int']
  balance: Scalars['Int']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  impUid?: Maybe<Scalars['ID']>
  status: Scalars['String']
  statusDetail: Scalars['String']
  updatedAt: Scalars['DateTime']
  useditem?: Maybe<IUseditem>
  user?: Maybe<IUser>
}

export type IQuery = {
  __typename?: 'Query'
  fetchBoard: IBoard
  fetchBoardComments: Array<IBoardComment>
  fetchBoards: Array<IBoard>
  fetchBoardsCount: Scalars['Int']
  fetchBoardsCountOfMine: Scalars['Int']
  fetchBoardsOfMine: Array<IBoard>
  fetchBoardsOfTheBest: Array<IBoard>
  fetchPointTransactions: Array<IPointTransaction>
  fetchPointTransactionsCountOfBuying: Scalars['Int']
  fetchPointTransactionsCountOfLoading: Scalars['Int']
  fetchPointTransactionsCountOfSelling: Scalars['Int']
  fetchPointTransactionsOfBuying: Array<IPointTransaction>
  fetchPointTransactionsOfLoading: Array<IPointTransaction>
  fetchPointTransactionsOfSelling: Array<IPointTransaction>
  fetchUseditem: IUseditem
  fetchUseditemQuestionAnswers: Array<IUseditemQuestionAnswer>
  fetchUseditemQuestions: Array<IUseditemQuestion>
  fetchUseditems: Array<IUseditem>
  fetchUseditemsCountIBought: Scalars['Int']
  fetchUseditemsCountIPicked: Scalars['Int']
  fetchUseditemsCountISold: Scalars['Int']
  fetchUseditemsIBought: Array<IUseditem>
  fetchUseditemsIPicked: Array<IUseditem>
  fetchUseditemsISold: Array<IUseditem>
  fetchUseditemsOfTheBest: Array<IUseditem>
  fetchUserLoggedIn: IUser
}

export type IQueryFetchBoardArgs = {
  boardId: Scalars['ID']
}
// ...
export type IToken = {
  __typename?: 'Token'
  accessToken: Scalars['String']
}
// ...
export type IUseditem = {
  __typename?: 'Useditem'
  _id: Scalars['ID']
  buyer?: Maybe<IUser>
  contents: Scalars['String']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  images?: Maybe<Array<Scalars['String']>>
  name: Scalars['String']
  pickedCount?: Maybe<Scalars['Int']>
  price?: Maybe<Scalars['Int']>
  remarks: Scalars['String']
  seller?: Maybe<IUser>
  soldAt?: Maybe<Scalars['DateTime']>
  tags?: Maybe<Array<Scalars['String']>>
  updatedAt: Scalars['DateTime']
  useditemAddress?: Maybe<IUseditemAddress>
}

export type IUseditemAddress = {
  __typename?: 'UseditemAddress'
  _id: Scalars['ID']
  address?: Maybe<Scalars['String']>
  addressDetail?: Maybe<Scalars['String']>
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  lat?: Maybe<Scalars['Float']>
  lng?: Maybe<Scalars['Float']>
  updatedAt: Scalars['DateTime']
  zipcode?: Maybe<Scalars['String']>
}

export type IUseditemAddressInput = {
  address?: InputMaybe<Scalars['String']>
  addressDetail?: InputMaybe<Scalars['String']>
  lat?: InputMaybe<Scalars['Float']>
  lng?: InputMaybe<Scalars['Float']>
  zipcode?: InputMaybe<Scalars['String']>
}

export type IUseditemQuestion = {
  __typename?: 'UseditemQuestion'
  _id: Scalars['ID']
  contents: Scalars['String']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  updatedAt: Scalars['DateTime']
  useditem: IUseditem
  user: IUser
}

export type IUseditemQuestionAnswer = {
  __typename?: 'UseditemQuestionAnswer'
  _id: Scalars['ID']
  contents: Scalars['String']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  updatedAt: Scalars['DateTime']
  useditemQuestion: IUseditemQuestion
  user: IUser
  fetchUseditemQuestionAnswers: any
}

export type IUser = {
  __typename?: 'User'
  _id: Scalars['ID']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  email: Scalars['String']
  name: Scalars['String']
  picture?: Maybe<Scalars['String']>
  updatedAt: Scalars['DateTime']
  userPoint?: Maybe<IUserPoint>
}

export type IUserPoint = {
  __typename?: 'UserPoint'
  _id: Scalars['ID']
  amount: Scalars['Int']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  updatedAt: Scalars['DateTime']
  user: IUser
}